syncdb/src/jp/co/ntt/oss/mapper/StructDataMapper.java
package jp.co.ntt.oss.mapper;

import java.sql.PreparedStatement;
import java.sql.ResultSet;
import java.sql.SQLException;
import java.sql.Types;

public class StructDataMapper implements DataMapper {

	@Override
	public final Object getObject(final ResultSet rset, final int columnIndex)
			throws SQLException {
		return rset.getObject(columnIndex);
	}

	@Override
	public final void setObject(final PreparedStatement pstmt,
			final int parameterIndex, final Object x) throws SQLException {
		pstmt.setObject(parameterIndex, x);
	}

	public static boolean isStructType(final int columnType) {
		if (columnType == Types.STRUCT) {
			return true;
		}

		return false;
	}
}
